feat(pbrAudioBlender): add new SpatialFrequencyResponseNode and rework utils/frd_io.py to add spatial frequency response

+++ properties/scenePG.py
@@ -69,7 +69,7 @@ class PBRAudioSceneProperties(PropertyGroup):
         update=set_quality
     )
 
-    sample_rate: IntProperty(
+    sample_rate: FloatProperty(
         name="Sample Rate",
         description="Audio sample rate in Hz",
         default=48000,
